still working on username modal
