+++ src/components/ApplicationView.js
@@ -5,7 +5,7 @@ import EventList from "./events/EventList";
 import LoginPage from "./auth/Auth";
 import Home from "./home/Home";
 import NewsArticleList from "./news/NewsArticleList";
-import SettingsList from "./settings/SettingsList";
+import SettingsCard from "./settings/SettingsCard";
 
 const ApplicationViews = props => {
   let history = useHistory();
@@ -59,7 +59,7 @@ const ApplicationViews = props => {
             sessionStorage.getItem("userId") === null ? (
               <Redirect exact to="/" />
             ) : (
-              <SettingsList {...props} />
+              <SettingsCard {...props} />
             )
           }
         ></Route>
